refactor : redux 폴더 내 hook과 type 폴더 / 파일 구분
